rename layout,compat as _layout and _compat
