+++ traitsui/testing/tester/_compat.py
