+++ traitsui/testing/tester/qt4/implementation/enum_editor.py
@@ -17,7 +17,7 @@ from traitsui.qt4.enum_editor import (
 from traitsui.testing.tester import command, locator, query
 from traitsui.testing.tester.common_ui_targets import _BaseSourceWithLocation
 from traitsui.testing.tester.qt4 import helpers
-from traitsui.testing.tester.editors.layout import column_major_to_row_major
+from traitsui.testing.tester.editors._layout import column_major_to_row_major
 
 
 class _IndexedListEditor(_BaseSourceWithLocation):
