+++ traitsui/testing/tester/wx/implementation/enum_editor.py
@@ -18,7 +18,7 @@ from traitsui.wx.enum_editor import (
 from traitsui.testing.tester import command, locator, query
 from traitsui.testing.tester.common_ui_targets import _BaseSourceWithLocation
 from traitsui.testing.tester.wx import helpers
-from traitsui.testing.tester.editors.layout import column_major_to_row_major
+from traitsui.testing.tester.editors._layout import column_major_to_row_major
 
 
 class _IndexedListEditor(_BaseSourceWithLocation):
